+++ hni-user/src/main/java/org/hni/user/service/UserPartialCreateService.java
@@ -4,5 +4,5 @@ import org.hni.common.service.BaseService;
 import org.hni.user.om.UserPartialData;
 
 public interface UserPartialCreateService extends BaseService<UserPartialData>{
-	public UserPartialData getUserPartialDataByUserId(int userId);
+	public UserPartialData getUserPartialDataByUserId(Long userId);
 }
